requirements.txt
azure-functions
azure-servicebus==7.14.0
azure-cosmos==4.3.0
requests
azure-storage-blob
azure-data-tables
